Move the priority helpers to the shared guide components and add a guide
- Move linearScore, the threshold raters, castEfficiencyPriority, activeTimePriority, and a generic resourceWastePriority into interface/guide/components/ImprovementPriorities
- Keep a thin Soul Shard wrapper for the Warlock specs
- Add README.md with the steps to add the section to a spec and to calibrate it
